adapt how we read in metadata from on-disk ipynb
